+++ src/pages/owner/ground/IndexPage.js
@@ -1,19 +1,12 @@
 import { useEffect } from "react";
 import { Outlet } from "react-router-dom";
-import BasicLayout from "layouts/BasicLayout";
+import BasicLayout from "layouts/OwnerLayout";
 const GroundIndexPage = () => {
-    useEffect(() => {
-
-    })
-    return (
-        <>
-
-           <BasicLayout>
-           <Outlet/>
-            </BasicLayout>
-            
-
-        </>
-    );
-}
+  useEffect(() => {});
+  return (
+      <BasicLayout>
+        <Outlet />
+      </BasicLayout>
+  );
+};
 export default GroundIndexPage;
