+++ src/pages/reserve/user/UserReservationStatusPage.js
@@ -3,7 +3,7 @@ import axios from "axios";
 import { Link, useNavigate } from "react-router-dom";
 
 import { getUserReservationStatus,reserveInfo } from "api/reserveApi";
-import PageComponent from "components/common/pageComponent";
+import PageComponent from "components/common/PageComponent";
 import { useSelector } from "react-redux";
 import BasicLayout from "layouts/BasicLayout";
